rename: Payment 폴더 domain 하위로 이동
