+++ src/main/java/site/goldenticket/common/security/SecurityConfiguration.java
@@ -39,6 +39,7 @@ public class SecurityConfiguration {
 
     private static final String[] PERMIT_ALL_URLS = new String[]{
             "/h2-console/**",
+            "/payments/**"
     };
 
     private static final String[] PERMIT_ALL_GET_URLS = new String[]{
