+++ src/main/java/site/goldenticket/domain/payment/controller/PaymentController.java
@@ -1,15 +1,15 @@
-package site.goldenticket.payment.controller;
+package site.goldenticket.domain.payment.controller;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import site.goldenticket.common.response.CommonResponse;
-import site.goldenticket.payment.dto.request.PaymentRequest;
-import site.goldenticket.payment.dto.response.PaymentDetailResponse;
-import site.goldenticket.payment.dto.response.PaymentReadyResponse;
-import site.goldenticket.payment.dto.response.PaymentResponse;
-import site.goldenticket.payment.service.PaymentService;
+import site.goldenticket.domain.payment.dto.response.PaymentDetailResponse;
+import site.goldenticket.domain.payment.dto.response.PaymentReadyResponse;
+import site.goldenticket.domain.payment.dto.request.PaymentRequest;
+import site.goldenticket.domain.payment.dto.response.PaymentResponse;
+import site.goldenticket.domain.payment.service.PaymentService;
 
 @RestController
 @RequestMapping("/payments")
@@ -31,7 +31,7 @@ public class PaymentController {
     }
 
     @PostMapping("/result")
-    public ResponseEntity<CommonResponse<PaymentResponse>> savePayment(@Valid PaymentRequest request) {
+    public ResponseEntity<CommonResponse<PaymentResponse>> savePayment(@Valid @RequestBody PaymentRequest request) {
         PaymentResponse paymentResponse = paymentService.savePayment(request);
         return ResponseEntity.ok(CommonResponse.ok("Payment result", paymentResponse));
     }
