+++ src/main/java/site/goldenticket/domain/payment/dto/response/PaymentReadyResponse.java
@@ -1,7 +1,7 @@
-package site.goldenticket.payment.dto.response;
+package site.goldenticket.domain.payment.dto.response;
 
-import site.goldenticket.payment.model.Order;
-import site.goldenticket.payment.service.PaymentService;
+import site.goldenticket.domain.payment.model.Order;
+import site.goldenticket.domain.payment.service.PaymentService;
 
 public record PaymentReadyResponse(
         Long id,
