+++ src/main/java/site/goldenticket/domain/payment/repository/IamportRepository.java
@@ -1,14 +1,15 @@
-package site.goldenticket.payment.repository;
+package site.goldenticket.domain.payment.repository;
 
 import com.siot.IamportRestClient.IamportClient;
+import com.siot.IamportRestClient.request.CancelData;
 import com.siot.IamportRestClient.request.PrepareData;
 import com.siot.IamportRestClient.response.IamportResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 import site.goldenticket.common.exception.CustomException;
 import site.goldenticket.common.response.ErrorCode;
-import site.goldenticket.payment.model.Payment;
-import site.goldenticket.payment.model.PaymentCancelDetail;
+import site.goldenticket.domain.payment.model.Payment;
+import site.goldenticket.domain.payment.model.PaymentCancelDetail;
 
 import java.math.BigDecimal;
 import java.util.List;
@@ -48,4 +49,13 @@ public class IamportRepository {
         }
     }
 
+    public List<PaymentCancelDetail> cancelPaymentByImpUid(String impUid) {
+        try {
+            IamportResponse<com.siot.IamportRestClient.response.Payment> response =
+                    iamportClient.cancelPaymentByImpUid(new CancelData(impUid, true));
+            return paymentCancelMapper.mapFrom(response.getResponse().getCancelHistory());
+        } catch (Exception e) {
+            throw new CustomException(ErrorCode.IAMPORT_ERROR);
+        }
+    }
 }
