+++ src/main/java/site/goldenticket/domain/payment/service/PaymentService.java
@@ -1,21 +1,20 @@
-package site.goldenticket.payment.service;
+package site.goldenticket.domain.payment.service;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import site.goldenticket.common.constants.OrderStatus;
-import site.goldenticket.common.constants.PaymentStatus;
 import site.goldenticket.common.exception.CustomException;
 import site.goldenticket.common.response.ErrorCode;
-import site.goldenticket.payment.dto.request.PaymentRequest;
-import site.goldenticket.payment.dto.response.PaymentDetailResponse;
-import site.goldenticket.payment.dto.response.PaymentReadyResponse;
-import site.goldenticket.payment.dto.response.PaymentResponse;
-import site.goldenticket.payment.model.Order;
-import site.goldenticket.payment.model.Payment;
-import site.goldenticket.payment.repository.IamportRepository;
-import site.goldenticket.payment.repository.OrderRepository;
-import site.goldenticket.payment.repository.PaymentRepository;
+import site.goldenticket.domain.payment.dto.response.PaymentDetailResponse;
+import site.goldenticket.domain.payment.dto.response.PaymentReadyResponse;
+import site.goldenticket.domain.payment.repository.IamportRepository;
+import site.goldenticket.domain.payment.repository.OrderRepository;
+import site.goldenticket.domain.payment.repository.PaymentRepository;
+import site.goldenticket.domain.payment.dto.request.PaymentRequest;
+import site.goldenticket.domain.payment.dto.response.PaymentResponse;
+import site.goldenticket.domain.payment.model.Order;
+import site.goldenticket.domain.payment.model.Payment;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
@@ -54,7 +53,7 @@ public class PaymentService {
         User user = new User();
 
         //TODO: productId,userId 이용하여 해당 유저가 네고를 진행 하였는지 확인, 진행하였다면 상품 가격 바꾸기
-        Integer price = 1000;
+        Integer price = 100;
 
         //TODO: productId로 상품 테이블에서 상품 가져오기
         Product product = new Product();
@@ -104,7 +103,7 @@ public class PaymentService {
 
     @Getter
     public static class Product {
-        private Long productId = 1L;
+        private Long productId = 100L;
         private Long userId = 101L;
         private String imageUrl = "default-image-url.jpg";
         private String accommodationName = "Default Accommodation";
